+++ gateway/src/apicast/policy/retry/retry.lua
@@ -1,0 +1,20 @@
+--- Retry policy
+
+local tonumber = tonumber
+
+local _M = require('apicast.policy').new('Retry Policy')
+
+local new = _M.new
+
+
+function _M.new(config)
+  local self = new(config)
+  self.retries = tonumber(config.retries)
+  return self
+end
+
+function _M:export()
+  return { upstream_retries = self.retries }
+end
+
+return _M