+++ gateway/src/resty/balancer.lua
@@ -149,4 +149,14 @@ function _M:set_timeouts(connect_timeout, send_timeout, read_timeout)
   return ngx_balancer.set_timeouts(connect_timeout, send_timeout, read_timeout)
 end
 
+function _M:retry_next_request()
+  local ngx_balancer = self.balancer
+
+  if not ngx_balancer then
+    return nil, 'balancer not available'
+  end
+
+  return ngx_balancer.set_more_tries(1)
+end
+
 return _M